Renaming of comonents
